-aChange unitcon_test.c

+++ src/unitcon.h
@@ -1,4 +1,4 @@
 #ifndef UNION_
 #define UNION_
-int Result (float number, char str1[], char str2[], char filename[]);
+double Result (float number, char str1[], char str2[], char filename[]);
 #endif
